Remove tgid check in fault handler
Remove the DBUG check in the page fault handler. For NVIDIA grace-
hopper nodes (at least) there is a separate process that handles
page faults and it is not in the original process' thread group.

Additionally, remove a redundant XPMEM_CMD_MAKE ioctl in xpmem_make.
That just creates a segment that is never used.

Synced from HPE xpmem community repo