Fix compute node crash in xpmem_vaddr_to_pte
A possible race between a process(that created xpmem memory region) trying to
migrate the Transparent Huge Page(THP) pages and another process(that attach
to that memory region) faulting on the same THP pages may cause a system crash.
To handle the race, the code has been enhanced accordingly.