Add additional sequence for per-VMA locks
The xpmem_detach() and xpmem_detach_att() perform a write-lock
on the struct mm_struct pointer passed to them that then in turn
performs a write to the VMA structure associated with it. Though
there are mutex locks taken for the struct xpmem_attachment pointer,
this does not satisfy the requirement for writes to the VMA while
the mmap_write_lock(mm) is held.

Test Description
- Ran xpmem_master test suite with no issues hit.