Four nodes rcu_sched stall during out of memory
While xpmem_fault_handler() is trying to pin a page, by calling
remap_pfn_range(), some other tasks may try to invalidate the same xpmem
page before it is completely pinned. This could lead to pinned memory, leak,
where a page was pinned, but was invalidated by xpmem_invalidate_page().
To prevent this, a call to acquire the invalidate_mutex of the
xpmem_attachment structure while calling remap_pfn_range() was added to make
sure the page can’t be invalidated until after it has been pinned.

Under extremely low memory condition, memory allocation in the process of
pinning a page may fail.  In this situation, the allocator would resort to
the slow path allocation, trying to free some previously allocated
memory to satisfy the allocation.  If the memory to be free is also xpmem
attached memory, it would result in a call to xpmem_invalidate_page(),
which also tries to acquire the invalidate_mutex of the xpmem_attachment
structure.

This could lead to a deadlock under the following 2 circumstances.

o The same mutex is being re-acquired again by the same process.
o In the normal xpmem_fault_handler() path, a process A could be holding
  an invalidate_mutex and attempting to acquire a page table lock.
  If the memory allocation failure mentioned above happens to another
  process B.  Process B should be holding invalidate_mutex, a page
  table lock, and trying to acquire another invalidate_mutex.
  We could run into a deadlock due to circular lock dependencies here, if
  process A, is holding mutex A and trying to lock spin_lock B (holding by
  process B), while process B holding mutex B and spin_lock B and trying to
  acquire mutex A (held by A).

To fix the issue, before calling remap_pfn_range() in xpmem_fault_handler()
a call to memalloc_noreclaim_save() is made to make sure that any future
memory allocation attempt by the process will fail if the kernel runs out of
memory.  There would be no attempt to do the slow path allocation, which
could lead to a deadlock seen here.