xpmem_mmu_release: hold lock across test and set of DESTROYING flag (#14)
If we release the spinlock after finding the flag not set, someone
else can come along and set it before we make a second check to see it
not set.

Change xpmem_teardown()'s interface so that the path that calls
xpmem_teardown() is responsible for setting the DESTROYING flag if
it's not already set - atomically - and calling xpmem_teardown() only
if it did so.

Also in the "self" case, do the same atomic test-and-set of the tg flag,
and don't call xpmem_teardown() if someone else had already set it.

Backported from hpc-shasta-os-xpmem repo commit: eef43d6a

Test Description:
Built on RHEL 9.6 and SLE 16 SP7 vm.
Ran xpmem functional tests without any test failures.
Verified the code path is exercised.